[Refactor] Move device commands and accessors out of components into library
Extract all of the code in *Status.react.js and DeviceListItem.react.js
for reading from redux and sending commands into nice classes, and make stuff
reusable with inheritance. These new files for each device are also a nice
place to put code for handling notification actions in the future.